+++ config/routes.rb
@@ -8,7 +8,14 @@ Rails.application.routes.draw do
     collection do
       get :history
     end
+    member do
+      get :correct
+    end
     resources :submissions, only: %i[new create show] do
+      member do
+        get :download
+      end
+
       resource :complete, only: :create, controller: 'submission_completion'
     end
     resource :template, only: %i[show]